requirements.txt
aiogram>=3.4,<4
fastapi>=0.110,<1
uvicorn[standard]>=0.23,<1
pydantic>=2.6,<3
pydantic-settings>=2.2,<3
sqlalchemy>=2.0,<2.1
alembic>=1.13,<2
asyncpg>=0.29,<1
redis>=5,<6
celery>=5.3,<6
python-dotenv>=1,<2
httpx>=0.27,<1
pillow>=10,<11
structlog>=24,<25
openai>=1.40,<2
anthropic>=0.31,<1
google-generativeai>=0.7,<0.8
boto3>=1.34,<2
minio>=7.2,<8
python-multipart>=0.0.9,<1
tenacity>=8.2,<9
typing-extensions>=4.8,<5


